pyproject.toml
[project]
name = "medical-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.6",
    "langchain-text-splitters>=1.1.0",
    "pydantic>=2.12.5",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
]
